rename to align to neutronjsplus changes
