+++ src/testcases/parallel/tokenfactory.test.ts
@@ -1,10 +1,10 @@
 import '@neutron-org/neutronjsplus';
 import {
   CosmosWrapper,
-  NEUTRON_DENOM,
   getEventAttribute,
 } from '@neutron-org/neutronjsplus/dist/cosmos';
-import { LocalState } from './../../helpers/localState';
+import { NEUTRON_DENOM } from '@neutron-org/neutronjsplus';
+import { LocalState, createWalletWrapper } from '../../helpers/localState';
 import { NeutronContract, Wallet } from '@neutron-org/neutronjsplus/dist/types';
 import {
   msgBurn,
@@ -17,10 +17,7 @@ import {
   getAuthorityMetadata,
   getBeforeSendHook,
 } from '@neutron-org/neutronjsplus/dist/tokenfactory';
-import {
-  WalletWrapper,
-  createWalletWrapper,
-} from '@neutron-org/neutronjsplus/dist/wallet_wrapper';
+import { WalletWrapper } from '@neutron-org/neutronjsplus/dist/walletWrapper';
 import { Suite, inject } from 'vitest';
 
 const config = require('../../config.json');
