+++ src/testcases/run_in_band/dex_bindings.test.ts
@@ -3,9 +3,9 @@ import {
   CosmosWrapper,
   getEventAttribute,
   getEventAttributesFromTx,
-  NEUTRON_DENOM,
 } from '@neutron-org/neutronjsplus/dist/cosmos';
-import { LocalState } from './../../helpers/localState';
+import { NEUTRON_DENOM } from '@neutron-org/neutronjsplus';
+import { LocalState, createWalletWrapper } from '../../helpers/localState';
 import { NeutronContract, CodeId } from '@neutron-org/neutronjsplus/dist/types';
 import {
   AllInactiveLimitOrderTrancheResponse,
@@ -25,15 +25,12 @@ import {
   PoolMetadataResponse,
   PoolReservesResponse,
   PoolResponse,
-} from '@neutron-org/neutronjsplus/dist/dex_bindings';
+} from '@neutron-org/neutronjsplus/dist/dexBindings';
 import {
   msgCreateDenom,
   msgMintDenom,
 } from '@neutron-org/neutronjsplus/dist/tokenfactory';
-import {
-  createWalletWrapper,
-  WalletWrapper,
-} from '@neutron-org/neutronjsplus/dist/wallet_wrapper';
+import { WalletWrapper } from '@neutron-org/neutronjsplus/dist/walletWrapper';
 
 const config = require('../../config.json');
